feat: add cli stylelint commands instead of creating scripts in package.json
