+++ test/commands/fix-all.test.ts
@@ -1,0 +1,17 @@
+import {expect, test} from '@oclif/test'
+
+describe('fix-all', () => {
+  test
+  .stdout()
+  .command(['fix-all'])
+  .it('runs hello', ctx => {
+    expect(ctx.stdout).to.contain('hello world')
+  })
+
+  test
+  .stdout()
+  .command(['fix-all', '--name', 'jeff'])
+  .it('runs hello --name jeff', ctx => {
+    expect(ctx.stdout).to.contain('hello jeff')
+  })
+})
