+++ test/commands/list.test.ts
@@ -1,0 +1,17 @@
+import {expect, test} from '@oclif/test'
+
+describe('list', () => {
+  test
+  .stdout()
+  .command(['list'])
+  .it('runs hello', ctx => {
+    expect(ctx.stdout).to.contain('hello world')
+  })
+
+  test
+  .stdout()
+  .command(['list', '--name', 'jeff'])
+  .it('runs hello --name jeff', ctx => {
+    expect(ctx.stdout).to.contain('hello jeff')
+  })
+})
